# Replace the Go Batch implementation with one based on RocksDB WriteBatch.