--- a/storage/raftstorage.go
+++ b/storage/raftstorage.go
@@ -280,6 +280,8 @@ func (r *Range) Snapshot() (raftpb.Snapshot, error) {
 // Append implements the multiraft.WriteableGroupStorage interface.
 func (r *Range) Append(entries []raftpb.Entry) error {
 	batch := r.rm.Engine().NewBatch()
+	defer batch.Close()
+
 	for _, ent := range entries {
 		err := engine.MVCCPutProto(batch, nil, engine.RaftLogKey(r.Desc().RaftID, ent.Index),
 			proto.ZeroTimestamp, nil, &ent)
@@ -312,7 +314,8 @@ func (r *Range) ApplySnapshot(snap raftpb.Snapshot) error {
 		return nil
 	}
 
-	batch := engine.NewBatch(r.rm.Engine())
+	batch := r.rm.Engine().NewBatch()
+	defer batch.Close()
 
 	// Delete everything in the range and recreate it from the snapshot.
 	for iter := newRangeDataIterator(r, r.rm.Engine()); iter.Valid(); iter.Next() {