Somewhat more involved than expected because the builtin RocksDB
WriteBatchWithIndex iterators do not handle merge records.

benchmark                         old ns/op     new ns/op     delta
BenchmarkMVCCBatch1Put10          6198          5223          -15.73%
BenchmarkMVCCBatch100Put10        6019          5586          -7.19%
BenchmarkMVCCBatch10000Put10      9324          5539          -40.59%
BenchmarkMVCCBatch100000Put10     13419         6242          -53.48%

benchmark                         old MB/s     new MB/s     speedup
BenchmarkMVCCBatch1Put10          1.61         1.91         1.19x
BenchmarkMVCCBatch100Put10        1.66         1.79         1.08x
BenchmarkMVCCBatch10000Put10      1.07         1.81         1.69x
BenchmarkMVCCBatch100000Put10     0.75         1.60         2.13x